implement remi suggested steps as Ansible tasks
When this role is invoked with a different PHP version than
previously installed it would fail at the dnf module enable step.
However it also provides a detailed message with the steps required.

```
The operation would result in switching of module 'php' stream 'remi-7.3' to stream
'remi-7.4' Error: It is not possible to switch enabled streams of a module. It is
recommended to remove all installed content from the module, and reset the module
using 'dnf module reset <module_name>' command. After you reset the module, you
can install the other stream
```

Resolves #49